cover: check for broken pipe
If cover was closed using an mpv keybind (such as pressing q on the mpv window)
instead of with a SIGINT (i.e. Ctrl-C) on the terminal, the metadata printer
remained active in the background, messing up future terminal output (it still
had the terminal's stdout open). Fix this by waiting not only for a metadata-
update event but also for the parent's death.