cover: find cover paths relative to working directory of audio player
I didn't notice this problem before because I was filtering everything through
an 'xargs realpath' in 'play'. Now that play doesn't do that, covers weren't
loading correctly. To fix this problem (& generalize further to an mpv launched
in another way), calculate cover paths relative to the working directory of the
mpv instance actually playing the media.