play: don't ignore symlinked media
Turns out GNU file(1) has non-POSIX behavior in that it doesn't follow symlinks
by default like it should, so symlinks reported a MIME type of inode/symlink
instead of that of the underlying file.

The JSON command generation was also improved by only spawning one jq instance
and streaming line-by-line. This is now the only part of the pipeline not null-
delimited. Finally, invocations of file after the initial one are nice'd, since
once something is playing there'll be a few minutes before the next media are
needed.