Replace which with command -v where possible
The which command is not POSIX so it's out. There are a few places I
use `which -a` to list all instances in $PATH instead of the one that
takes precendece/would be actually executed, and I'm too lazy to write
a new script that does that manually, so those remain.