launch-portals: add new script
OpenRC does not have a parallel to systemd's user units, so we have to
use our own poor imitation. This script launches the xdg-desktop-portal
and pipewire stuff necessary for screensharing on Sway with Chromium as
of 2023-05-06. (It should also be sufficient for pipewire[sound-server]
as long as you don't need pipewire-pulse, which I hope not to.)