footclient: add new script
When running the foot terminal emulator in "server mode" (in which a
daemon updates the wl_surfaces for potentially many client terminals),
the foot server can take a second or two to start (or even stop itself
if your config is broken, etc.) which would leave you terminal-less in
an environment where your "open terminal" keybind starts a foot client
and not a standalone foot terminal. Wrapping footclient in this script
makes it wait for the server socket to become visible, effectively
queueing "open terminal" requests so you don't wait an indeterminate
time for an invisible process to come alive.