flacless: add new script
I got a new phone that doesn't have enough storage to dedicate 156 GB
to music, so until I make IRL Pied Piper we will have to make do with
256 kbps AAC which is almost transparent/perceptually lossless. Also,
if I stream via Bluetooth to AirPods, the audio stream is 256 kbps AAC
at best anyway. (Some app/OS/BT stack combos can even directly forward
encoded audio if it's of the right format, etc. which means choosing
this specific format and bit rate might obviate on-device re-encoding
as well. I'm not sure if this is the case for my current phone though.)

This script copies lossy file formats verbatim to a "flacless" folder,
but converts FLAC files to mp3s (with libfdk_aac at 256 kbps for the
reasons mentioned earlier). It caches flac->mp3 encodes and hardlinks
them into place, so the script can be run "incrementally" with much
lower runtime than if it started from scratch. (Even on a zillion-core
machine, when you have as much music and are as impatient as I am, this
matters!)

At the moment, this script assumes quite a lot about the environment,
such as >75GB storage at /raid0/flacless and a Syncthing folder titled
music-flacless. This may be fixed in a future version.