swaylock-transparent: add new binary
This is a build of the last swaylock commit before they removed
layer-shell support in favor of ext-session-lock-v1, which is an
improvement in every respect except support for transparent lock
screens, which don't work anymore. swaylock-transparent stubs out
ext-session-lock-v1 code so it's guaranteed to use layer-shell and
work with transparent background images or colors.

Specifically, swaylock-transparent is built from commit 3175439 at:
https://github.com/milkey-mouse/swaylock/tree/layer-shell-only